(fix) refactor server to use new organized queries
