+++ src/server/server.js
@@ -1,7 +1,7 @@
 var express = require('express');
 var http = require('http');
 var bodyParser = require('body-parser');
-var db = require('./queries');
+var db = require('./queries/exports');
 var path = require('path');
 
 var PORT = process.env.PORT || 3000;
@@ -37,13 +37,13 @@ app.post('/api/users/:authID', db.setAuthID);
 
 
 app.post('/api/teachers/:authID', db.insertTeacher);
+app.post('/api/available/:authID', db.setAvailability);
 
 
 
 app.put('/api/users/:authID', db.updateUser);
 app.delete('/api/users/:authID', db.removeUser);
 app.delete('/api/subject/:id', db.removeSubject);
-app.post('/api/available/:authID', db.setAvailability);
 app.get('/api/learning/:id', db.findSubjectsByUser);
 app.delete('/api/learning/:userID/:subjectID', db.removeSubjectByUser);
